Move Snippets main to function
