+++ packages/snippets/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log for `snippets`
 
+## 0.1.1
+
+* Moves the snippets main into a function that can be called from the framework (`snippetsMain`).
+
 ## 0.1.0
 
 * Initial version, ported and modified from the [snippets tool](https://github.com/flutter/flutter/tree/master/dev/snippets) in the Flutter repo.
