+++ packages/snippets/bin/snippets.dart
@@ -2,204 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:io' show Process, ProcessResult, Platform, stderr, exit;
-
-import 'package:args/args.dart';
-import 'package:file/file.dart';
-import 'package:file/local.dart';
-import 'package:path/path.dart' as path;
-import 'package:snippets/snippets.dart';
-
-const String _kSerialOption = 'serial';
-const String _kElementOption = 'element';
-const String _kHelpOption = 'help';
-const String _kInputOption = 'input';
-const String _kLibraryOption = 'library';
-const String _kOutputOption = 'output';
-const String _kPackageOption = 'package';
-const String _kTemplateOption = 'template';
-const String _kTypeOption = 'type';
-
-const LocalFileSystem filesystem = LocalFileSystem();
-
-String getChannelName() {
-  final RegExp gitBranchRegexp = RegExp(r'^## (?<branch>.*)');
-  final ProcessResult gitResult =
-      Process.runSync('git', <String>['status', '-b', '--porcelain']);
-  if (gitResult.exitCode != 0)
-    throw 'git status exit with non-zero exit code: ${gitResult.exitCode}';
-  final RegExpMatch? gitBranchMatch = gitBranchRegexp
-      .firstMatch((gitResult.stdout as String).trim().split('\n').first);
-  return gitBranchMatch == null
-      ? '<unknown>'
-      : gitBranchMatch.namedGroup('branch')!.split('...').first;
-}
-
-const List<String> sampleTypes = <String>[
-  'snippet',
-  'sample',
-  'dartpad',
-];
+import 'dart:io';
+import 'package:snippets/publish.dart';
 
 /// Generates snippet dartdoc output for a given input, and creates any sample
 /// applications needed by the snippet.
 void main(List<String> argList) {
-  final Map<String, String> environment = Platform.environment;
-  final ArgParser parser = ArgParser();
-  parser.addOption(
-    _kTypeOption,
-    defaultsTo: 'dartpad',
-    allowed: sampleTypes,
-    allowedHelp: <String, String>{
-      'dartpad':
-          'Produce a code sample application complete with embedding the sample in an '
-              'application template for using in Dartpad.',
-      'sample':
-          'Produce a code sample application complete with embedding the sample in an '
-              'application template.',
-      'snippet':
-          'Produce a nicely formatted piece of sample code. Does not embed the '
-              'sample into an application template.',
-    },
-    help: 'The type of snippet to produce.',
-  );
-  parser.addOption(
-    _kTemplateOption,
-    defaultsTo: null,
-    help: 'The name of the template to inject the code into.',
-  );
-  parser.addOption(
-    _kOutputOption,
-    defaultsTo: null,
-    help: 'The output path for the generated sample application. Overrides '
-        'the naming generated by the --package/--library/--element arguments. '
-        'Metadata will be written alongside in a .json file. '
-        'The basename of this argument is used as the ID',
-  );
-  parser.addOption(
-    _kInputOption,
-    defaultsTo: environment['INPUT'],
-    help: 'The input file containing the sample code to inject.',
-  );
-  parser.addOption(
-    _kPackageOption,
-    defaultsTo: environment['PACKAGE_NAME'],
-    help: 'The name of the package that this sample belongs to.',
-  );
-  parser.addOption(
-    _kLibraryOption,
-    defaultsTo: environment['LIBRARY_NAME'],
-    help: 'The name of the library that this sample belongs to.',
-  );
-  parser.addOption(
-    _kElementOption,
-    defaultsTo: environment['ELEMENT_NAME'],
-    help: 'The name of the element that this sample belongs to.',
-  );
-  parser.addOption(
-    _kSerialOption,
-    defaultsTo: environment['INVOCATION_INDEX'],
-    help: 'A unique serial number for this snippet tool invocation.',
-  );
-  parser.addFlag(
-    _kHelpOption,
-    defaultsTo: false,
-    negatable: false,
-    help: 'Prints help documentation for this command',
-  );
-
-  final ArgResults args = parser.parse(argList);
-
-  if (args[_kHelpOption] as bool) {
-    stderr.writeln(parser.usage);
-    exit(0);
-  }
-
-  final String sampleType = args[_kTypeOption] as String;
-
-  if (args[_kInputOption] == null) {
-    stderr.writeln(parser.usage);
-    errorExit(
-        'The --$_kInputOption option must be specified, either on the command '
-        'line, or in the INPUT environment variable.');
-  }
-
-  final File input = filesystem.file(args['input'] as String);
-  if (!input.existsSync()) {
-    errorExit('The input file ${input.path} does not exist.');
-  }
-
-  String? template;
-  if (sampleType == 'sample' || sampleType == 'dartpad') {
-    final String templateArg = args[_kTemplateOption] as String;
-    if (templateArg.isEmpty) {
-      stderr.writeln(parser.usage);
-      errorExit(
-          'The --$_kTemplateOption option must be specified on the command '
-          'line for application samples.');
-    }
-    template = templateArg.replaceAll(RegExp(r'.tmpl$'), '');
-  }
-
-  final String packageName = args[_kPackageOption] as String? ?? '';
-  final String libraryName = args[_kLibraryOption] as String? ?? '';
-  final String elementName = args[_kElementOption] as String? ?? '';
-  final String serial = args[_kSerialOption] as String? ?? '';
-  final List<String> id = <String>[];
-  File? output;
-  if (args[_kOutputOption] != null) {
-    id.add(path.basename(
-        path.basenameWithoutExtension(args[_kOutputOption] as String)));
-    output = filesystem.file(path.absolute(args[_kOutputOption] as String));
-  } else {
-    if (packageName.isNotEmpty && packageName != 'flutter') {
-      id.add(packageName);
-    }
-    if (libraryName.isNotEmpty) {
-      id.add(libraryName);
-    }
-    if (elementName.isNotEmpty) {
-      id.add(elementName);
-    }
-    if (serial.isNotEmpty) {
-      id.add(serial);
-    }
-    if (id.isEmpty) {
-      errorExit('Unable to determine ID. At least one of --$_kPackageOption, '
-          '--$_kLibraryOption, --$_kElementOption, -$_kSerialOption, or the environment variables '
-          'PACKAGE_NAME, LIBRARY_NAME, ELEMENT_NAME, or INVOCATION_INDEX must be non-empty.');
-    }
-  }
-
-  final int? sourceLine = environment['SOURCE_LINE'] != null
-      ? int.tryParse(environment['SOURCE_LINE']!)
-      : null;
-  final String sourcePath = environment['SOURCE_PATH'] ?? 'unknown.dart';
-  final SnippetDartdocParser sampleParser = SnippetDartdocParser();
-  final SourceElement element = sampleParser.parseFromDartdocToolFile(
-    input,
-    startLine: sourceLine,
-    element: elementName,
-    sourceFile: filesystem.file(sourcePath),
-    template: template!,
-    type: sampleType,
-  );
-  final SnippetGenerator generator = SnippetGenerator();
-  final Map<String, Object?> metadata = <String, Object?>{
-    'channel': getChannelName(),
-    'serial': serial,
-    'package': packageName,
-    'library': libraryName,
-    'element': elementName,
-  };
-  for (final CodeSample sample in element.samples) {
-    sample.metadata.addAll(metadata);
-    generator.generateCode(
-      sample,
-      output: output,
-    );
-    print(generator.generateHtml(sample));
-  }
-
+  snippetMain(argList);
   exit(0);
 }
